+++ backend/src/main/java/io/metersphere/api/service/ApiTestCaseService.java
@@ -93,6 +93,11 @@ public class ApiTestCaseService {
 
     public List<ApiTestCaseResult> list(ApiTestCaseRequest request) {
         request.setOrders(ServiceUtils.getDefaultOrder(request.getOrders()));
+        if (request.getModuleIds() == null) {
+            List<String> moduleIds = new ArrayList<>();
+            moduleIds.add(request.getModuleId());
+            request.setModuleIds(moduleIds);
+        }
         List<ApiTestCaseResult> returnList = extApiTestCaseMapper.list(request);
 
         for (ApiTestCaseResult res : returnList) {
