+++ frontend/src/business/components/track/case/components/TestCaseEditOtherInfo.vue
@@ -99,7 +99,7 @@ import FormRichTextItem from "@/business/components/track/case/components/FormRi
 export default {
   name: "TestCaseEditOtherInfo",
   components: {FormRichTextItem, TestCaseIssueRelate, TestCaseAttachment, MsRichText, TestCaseRichText},
-  props: ['form', 'labelWidth', 'caseId', 'readOnly', 'projectId', 'isTestPlan', 'planId'],
+  props: ['form', 'labelWidth', 'caseId', 'readOnly', 'projectId', 'isTestPlan', 'planId', 'sysList'],
   data() {
     return {
       result: {},
@@ -108,7 +108,7 @@ export default {
       fileList: [],
       tableData: [],
       demandOptions: [],
-      sysList: [],//一级选择框的数据
+      //sysList:this.sysList,//一级选择框的数据
       props: {
         multiple: true,
         //lazy: true,
@@ -123,9 +123,7 @@ export default {
   },
   watch: {
     tabActiveName() {
-      if (this.tabActiveName === 'relateTest') {
-        this.loadOptions(this.sysList);
-      } else if (this.tabActiveName === 'demand') {
+      if (this.tabActiveName === 'demand') {
         this.getDemandOptions();
       } else if (this.tabActiveName === 'bug') {
         this.$refs.issue.getIssues();
@@ -267,90 +265,7 @@ export default {
         });
       }
     },
-    async loadOptions(sysLib) {
-      if (this.form.list) {
-        return;
-      }
 
-      sysLib = TEST
-        .filter(item => {
-          return enableModules([item.module]);
-        })// 模块启用禁用过滤
-        .map(item => ({
-          value: item.id,
-          label: item.name,
-        }));
-      let array = [];
-      for (let i = 0; i < sysLib.length; i++) {
-        if (sysLib.length > 0) {
-          let res = await this.getTestOptions(sysLib[i].value);
-          sysLib[i].children = res;
-        }
-        array.push(sysLib[i]);
-      }
-      this.sysList = array;
-    },
-    getTestOptions(val) {
-      this.form.type = val;
-      this.testOptions = [];
-      let url = '';
-      if (this.form.type === 'testcase') {
-        url = '/api/' + this.form.type + '/list/' + this.projectId;
-        if (!url) {
-          return;
-        }
-        this.buildValue(url);
-      } else if (this.form.type === 'performance' || this.form.type === 'api') {
-        url = '/' + this.form.type + '/list/' + this.projectId;
-        if (!url) {
-          return;
-        }
-        this.buildValue(url);
-      } else if (this.form.type === 'automation') {
-        //url = '/api/' + this.form.type + '/list/' + this.projectId;
-        url = '/api/automation/module/list/' + this.projectId;
-        if (!url) {
-          return;
-        }
-        this.result.loading = true;
-        return new Promise((resolve, reject) => {
-          this.$get("/api/automation/module/list/" + this.projectId, response => {
-            if (response.data != undefined && response.data != null) {
-              this.buildTreeValue(response.data);
-            }
-            this.result.loading = false;
-            resolve(response.data);
-          });
-        });
-      }
-    },
-
-    buildTreeValue(list) {
-      list.forEach(item => {
-        item.value = item.id,
-            item.label = item.name,
-            item.leaf = true;
-        if (item.children) {
-          this.buildTreeValue(item.children);
-        }
-      });
-    },
-    buildValue(url) {
-      this.result.loading = true;
-      return new Promise((resolve, reject) => {
-        this.$get(url).then(res => {
-          const data = res.data.data.map(item => ({
-            value: item.id,
-            label: item.name,
-            leaf: true
-          }));
-          this.result.loading = false;
-          resolve(data);
-        }).catch((err) => {
-          reject(err);
-        });
-      });
-    }
   }
 };
 </script>
@@ -371,6 +286,6 @@ export default {
 
 .el-cascader >>> .el-input {
   cursor: pointer;
-  width: 300px;
+  width: 500px;
 }
 </style>
